requirements.txt
python-dotenv
easyocr
openai
transformers